move components to corresponding files
