+++ packages/styleguide/src/components/Preview/Interact/Interact.js
@@ -1,76 +1,19 @@
 /* eslint-disable no-underscore-dangle */
-import React, { createContext } from 'react';
+import React from 'react';
 import { arrayOf, string, bool, element, func, oneOfType } from 'prop-types';
 import styled from 'styled-components';
 import CodeExample from '../CodeExample';
 
-import { Bar, BarItem } from '../../Bar';
-import Button from '../../Button';
-
 import useId from '../../../utils/useId';
 
-const InteractContext = createContext();
+import PropInput from './PropInput';
+import PropSelect from './PropSelect';
+import PropCheckbox from './PropCheckbox';
+import PropsGroup from './PropsGroup';
+import ButtonShowPropsGroup from './ButtonShowPropsGroup';
+import InteractContext from './state';
 
-/**
- * Cleanup docgen prop value of string which has this form:
- * "'string'" and converts it to "string"
- */
-const cleanValue = s =>
-  typeof s === 'string' ? s.replace(/(^['"]|['"]$)/g, '') : s;
-
-/**
- * Get docgen props of component
- * if it's an html element we parse just first level of text to be editable
- * if it's just a string we create docgen type of string
- */
-const getDocgenProps = component => {
-  const componentType = component.type;
-
-  if (typeof componentType === 'function') {
-    if (componentType.__docgenInfo) {
-      return componentType.__docgenInfo.props;
-    }
-  } else if (typeof componentType === 'string') {
-    const props = component.props;
-    return Object.keys(props).reduce(
-      (acc, curr) => ({
-        ...acc,
-        [curr]: {
-          type: {
-            name: typeof props[curr],
-            value: props[curr],
-          },
-        },
-      }),
-      {}
-    );
-  }
-
-  return {
-    children: {
-      type: {
-        name: typeof component,
-        value: component,
-      },
-    },
-  };
-};
-
-/**
- * parse displayName of component or create custom name
- */
-const getDisplayName = component =>
-  (component.type && component.type.displayName) ||
-  (typeof component === 'string'
-    ? 'PlainText'
-    : typeof component.type === 'string'
-    ? component.type
-    : 'NoDisplayName');
-
-/**
- * Helper function for comparision
- */
-const is = (value, type) => value === type;
+import { cleanValue, getDocgenProps, getDisplayName, is } from './helpers';
 
 /**
  * Class that creates interactive playground
@@ -586,206 +529,6 @@ class Interact extends React.Component {
   }
 }
 
-const PropLabelWithTooltip = ({ isDisabled, inputProps, componentInfo }) => (
-  <Bar space="tiny">
-    <BarItem>
-      <PropLabel
-        {...inputProps}
-        {...(isDisabled ? { style: { color: 'grey' } } : {})}
-      />
-    </BarItem>
-    <BarItem>
-      <PropTooltip {...componentInfo} />
-    </BarItem>
-  </Bar>
-);
-
-const PropLabel = ({ id, name, ...other }) => (
-  <label htmlFor={id} {...other}>
-    <strong>{name.replace(/^\w/, m => m.toUpperCase())}</strong>
-  </label>
-);
-
-const PropTooltip = ({ id, name }) => (
-  <InteractContext.Consumer>
-    {({ docgen }) => {
-      const propInfo = docgen.liveProps[id][name];
-      return (
-        <Tooltip
-          dialog={propInfo.description || 'No valuable description found'}
-        >
-          i
-        </Tooltip>
-      );
-    }}
-  </InteractContext.Consumer>
-);
-
-const PropInput = ({ type, isDisabled, inputProps, componentInfo }) => (
-  <InteractContext.Consumer>
-    {({ state, handleInputChange }) => {
-      const { id, name } = componentInfo;
-      let value = state.liveProps[id][name];
-
-      return (
-        <React.Fragment>
-          <PropLabelWithTooltip
-            isDisabled={isDisabled}
-            inputProps={inputProps}
-            componentInfo={componentInfo}
-          />
-          {isDisabled ? (
-            <Input
-              {...inputProps}
-              disabled
-              value={
-                is(typeof value, 'object')
-                  ? JSON.stringify(value)
-                  : value.toString()
-              }
-            />
-          ) : (
-            <Input
-              {...inputProps}
-              value={value}
-              onChange={e => {
-                handleInputChange(type, e);
-              }}
-            />
-          )}
-        </React.Fragment>
-      );
-    }}
-  </InteractContext.Consumer>
-);
-
-const PropCheckbox = ({ inputProps, componentInfo }) => {
-  const { id, name } = componentInfo;
-
-  return (
-    <InteractContext.Consumer>
-      {({ state, handleCheckboxChange }) => (
-        <Bar>
-          <BarItem>
-            <Input
-              {...inputProps}
-              type="checkbox"
-              checked={state.liveProps[id][name] ? 'checked' : false}
-              onChange={handleCheckboxChange}
-            />{' '}
-          </BarItem>
-          <BarItem>
-            <PropLabelWithTooltip
-              inputProps={inputProps}
-              componentInfo={componentInfo}
-            />
-          </BarItem>
-        </Bar>
-      )}
-    </InteractContext.Consumer>
-  );
-};
-
-const PropSelect = ({ inputProps, componentInfo }) => {
-  const { id, name } = componentInfo;
-
-  return (
-    <InteractContext.Consumer>
-      {({ state, docgen, handleSelectChange }) => {
-        const propInfo = docgen.liveProps[id][name];
-
-        return (
-          <React.Fragment>
-            <PropLabelWithTooltip
-              inputProps={inputProps}
-              componentInfo={componentInfo}
-            />
-            <select
-              {...inputProps}
-              value={state.liveProps[id][name]}
-              onChange={handleSelectChange}
-            >
-              <option value="null">None</option>
-              {Object.values(propInfo.type.value).map((option, index) => {
-                const optionValue = cleanValue(option.value);
-                return (
-                  <option key={index.toString()} value={optionValue}>
-                    {optionValue}
-                  </option>
-                );
-              })}
-            </select>
-          </React.Fragment>
-        );
-      }}
-    </InteractContext.Consumer>
-  );
-};
-
-const PropsGroup = ({ id }) => (
-  <InteractContext.Consumer>
-    {({ renderInput, state, props, docgen }) => {
-      const statePropNames = Object.keys(docgen.liveProps[id]);
-      const propCount = statePropNames.length;
-
-      return (
-        state.showProps[id] && (
-          <StyledWrapper>
-            {propCount ? (
-              statePropNames.map(name => {
-                return props.filterProps.find(filtered => filtered === name)
-                  ? null
-                  : renderInput(id, name);
-              })
-            ) : (
-              <p>There are no props to edit, try another component!</p>
-            )}
-          </StyledWrapper>
-        )
-      );
-    }}
-  </InteractContext.Consumer>
-);
-
-const ButtonShowPropsGroup = ({ id, componentName, deepness }) => (
-  <InteractContext.Consumer>
-    {({ handleShowProps, state }) => (
-      <Button
-        fontSize="base"
-        onClick={e => {
-          handleShowProps(e, id);
-        }}
-        style={{ paddingLeft: 0 }}
-      >
-        {deepness > 0 && '↳'} {componentName} {state.showProps[id] ? '▲' : '▼'}
-      </Button>
-    )}
-  </InteractContext.Consumer>
-);
-
-const Tooltip = ({ children, dialog, ...other }) => (
-  <StyledTooltip {...other}>
-    <div className="tooltip-trigger">{children}</div>
-    <div className="tooltip-title">{dialog}</div>
-  </StyledTooltip>
-);
-
-const StyledWrapper = styled.div`
-  border-left: 1px solid ${props => props.theme.colors.greyDark};
-  padding-left: ${props => props.theme.spaces.small};
-  position: relative;
-
-  &::before {
-    content: '';
-    position: absolute;
-    bottom: -20px;
-    left: -1px;
-    height: 20px;
-    width: 0;
-    border-left: 1px solid ${props => props.theme.colors.greyDark};
-  }
-`;
-
 const StyledInteract = styled.div`
   font-family: ${props => props.theme.fontFamily};
 `;
@@ -803,68 +546,6 @@ const GridCol = styled.div`
     `
     width: ${(props.size / 12) * 100 + '%'}
     `}
-`;
-
-const Input = styled.input`
-  background-color: ${props => props.theme.colors.white};
-  border: 1px solid ${props => props.theme.colors.greyDark};
-  border-radius: 3px;
-  font-size: 0.875rem;
-  min-height: 1em;
-  padding: 0.42857em 0.57143em;
-
-  &:disabled {
-    opacity: 0.4;
-    cursor: not-allowed;
-  }
-`;
-
-const StyledTooltip = styled.div`
-  position: relative;
-  color: ${props => props.theme.colors.white};
-
-  .tooltip-title {
-    pointer-events: none;
-    position: absolute;
-    width: 200px;
-    text-align: center;
-    padding: ${props => props.theme.spaces.small};
-    background: black;
-    border-radius: ${props => props.theme.spaces.tiny};
-    left: 0;
-    top: 50%;
-    transform: translate(-105%, -50%);
-    opacity: 0;
-    transition: opacity 0.3s ease-in 0.1s;
-
-    ::after {
-      content: '';
-      position: absolute;
-      right: 0;
-      top: 50%;
-      transform: translate(100%, -50%);
-      border-style: solid;
-      border-width: 6px 0 6px 8px;
-      border-color: transparent transparent transparent
-        ${props => props.theme.colors.black};
-    }
-  }
-
-  .tooltip-trigger:hover + .tooltip-title {
-    opacity: 1;
-  }
-
-  .tooltip-trigger {
-    background: ${props => props.theme.colors.black};
-    color: ${props => props.theme.colors.white};
-    text-align: center;
-    font-size: 16px;
-    line-height: 24px;
-    height: 24px;
-    width: 24px;
-    margin: 0;
-    border-radius: 50%;
-  }
 `;
 
 const StyledSticky = styled.div`
