+++ packages/styleguide/src/components/Preview/Interact/state.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const InteractContext = createContext();
+
+export default InteractContext;
